# api/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Tracker API
  description: Single-user gym tracking API. All endpoints except /api/auth/token require a bearer JWT.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
tags:
  - name: auth
    description: Token issuance
  - name: exercises
    description: Exercise catalog
  - name: workouts
    description: Workout sessions
  - name: sets
    description: Sets within a workout exercise
  - name: templates
    description: Workout templates

paths:
  /api/auth/token:
    post:
      tags: [auth]
      summary: Issue a JWT for the configured user
      operationId: issueToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/exercises:
    get:
      tags: [exercises]
      summary: List exercises
      operationId: listExercises
      parameters:
        - name: muscleGroup
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MuscleGroup'
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Exercise list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [exercises]
      summary: Create an exercise
      operationId: createExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseRequest'
      responses:
        '201':
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: An exercise with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/exercises/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [exercises]
      summary: Get an exercise
      operationId: getExercise
      responses:
        '200':
          description: Exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [exercises]
      summary: Update an exercise
      operationId: updateExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseRequest'
      responses:
        '200':
          description: Exercise updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [exercises]
      summary: Archive an exercise (soft delete)
      description: Sets archived=true. The exercise remains in historical workouts but no longer appears in the catalog by default.
      operationId: deleteExercise
      responses:
        '204':
          description: Exercise archived
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [exercises]
      summary: Partially update an exercise
      operationId: patchExercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchExerciseRequest'
      responses:
        '200':
          description: Exercise updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/workouts:
    post:
      tags: [workouts]
      summary: Start a new workout session
      operationId: createWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkoutRequest'
      responses:
        '201':
          description: Workout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [workouts]
      summary: List workouts, most recent first
      operationId: listWorkouts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated workout list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPage'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/workouts/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [workouts]
      summary: Get full workout detail with exercises and sets
      operationId: getWorkout
      responses:
        '200':
          description: Workout detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [workouts]
      summary: Update a workout (notes, finish time)
      operationId: updateWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutRequest'
      responses:
        '200':
          description: Workout updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [workouts]
      summary: Delete a workout and all its sets
      operationId: deleteWorkout
      responses:
        '204':
          description: Workout deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/workouts/{workoutId}/exercises/{exerciseId}/sets:
    parameters:
      - $ref: '#/components/parameters/WorkoutId'
      - $ref: '#/components/parameters/ExerciseId'
    post:
      tags: [sets]
      summary: Log a set for an exercise within a workout
      description: If the exercise is not yet part of the workout, it is added automatically at the next position.
      operationId: createSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSetRequest'
      responses:
        '201':
          description: Set logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/workouts/{workoutId}/exercises/{exerciseId}/sets/{setId}:
    parameters:
      - $ref: '#/components/parameters/WorkoutId'
      - $ref: '#/components/parameters/ExerciseId'
      - $ref: '#/components/parameters/SetId'
    put:
      tags: [sets]
      summary: Update a logged set
      operationId: updateSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSetRequest'
      responses:
        '200':
          description: Set updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [sets]
      summary: Delete a logged set
      operationId: deleteSet
      responses:
        '204':
          description: Set deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/templates:
    get:
      tags: [templates]
      summary: List templates
      operationId: listTemplates
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [templates]
      summary: Create a template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: A referenced exercise does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A template with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/templates/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [templates]
      summary: Get template detail with ordered exercises and targets
      operationId: getTemplate
      responses:
        '200':
          description: Template detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [templates]
      summary: Update a template (full replace of the exercise list)
      operationId: updateTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A template with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [templates]
      summary: Delete a template
      description: Hard delete. Workouts previously started from this template are kept (their templateId becomes null).
      operationId: deleteTemplate
      responses:
        '204':
          description: Template deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/templates/{id}/start:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [templates]
      summary: Start a workout from this template
      description: Creates a new workout pre-populated with the template's exercises in order (no sets). Later edits to the template do not affect the created workout.
      operationId: startWorkoutFromTemplate
      responses:
        '201':
          description: Workout created from template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    WorkoutId:
      name: workoutId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ExerciseId:
      name: exerciseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SetId:
      name: setId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    MuscleGroup:
      type: string
      enum:
        - CHEST
        - BACK
        - SHOULDERS
        - BICEPS
        - TRICEPS
        - FOREARMS
        - QUADS
        - HAMSTRINGS
        - GLUTES
        - CALVES
        - CORE
        - FULL_BODY

    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    AuthResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token lifetime in seconds
          example: 86400

    Exercise:
      type: object
      required: [id, name, muscleGroups, defaultRestSeconds, archived]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        muscleGroups:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MuscleGroup'
        defaultRestSeconds:
          type: integer
          minimum: 0
          example: 120
        notes:
          type: string
          nullable: true
        archived:
          type: boolean
          example: false

    CreateExerciseRequest:
      type: object
      required: [name, muscleGroups]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        muscleGroups:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MuscleGroup'
        defaultRestSeconds:
          type: integer
          minimum: 0
          default: 120
        notes:
          type: string
          nullable: true
          maxLength: 500

    WorkoutSet:
      type: object
      required: [id, setNumber, reps, weightKg, completedAt]
      properties:
        id:
          type: integer
          format: int64
        setNumber:
          type: integer
          minimum: 1
        reps:
          type: integer
          minimum: 1
        weightKg:
          type: number
          format: double
          minimum: 0
        rpe:
          type: number
          format: double
          minimum: 1
          maximum: 10
          nullable: true
        completedAt:
          type: string
          format: date-time

    CreateSetRequest:
      type: object
      required: [reps, weightKg]
      properties:
        reps:
          type: integer
          minimum: 1
        weightKg:
          type: number
          format: double
          minimum: 0
        rpe:
          type: number
          format: double
          minimum: 1
          maximum: 10
          nullable: true

    WorkoutExercise:
      type: object
      required: [exercise, position, sets]
      properties:
        exercise:
          $ref: '#/components/schemas/Exercise'
        position:
          type: integer
          minimum: 1
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSet'

    Workout:
      type: object
      required: [id, startedAt, exercises]
      properties:
        id:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExercise'

    CreateWorkoutRequest:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
          description: Defaults to server time when omitted
        notes:
          type: string
          nullable: true
          maxLength: 1000

    UpdateWorkoutRequest:
      type: object
      properties:
        finishedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
          maxLength: 1000

    WorkoutSummary:
      type: object
      required: [id, startedAt, exerciseCount, totalSets, totalVolumeKg]
      properties:
        id:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        exerciseCount:
          type: integer
        totalSets:
          type: integer
        totalVolumeKg:
          type: number
          format: double

    WorkoutPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSummary'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    ErrorResponse:
      type: object
      required: [status, error, message, timestamp]
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          nullable: true
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string

    PatchExerciseRequest:
      type: object
      properties:
        defaultRestSeconds:
          type: integer
          minimum: 0

    TemplateExerciseItem:
      type: object
      required: [exercise, position, targetSets, targetReps]
      properties:
        exercise:
          $ref: '#/components/schemas/Exercise'
        position:
          type: integer
          minimum: 1
        targetSets:
          type: integer
          minimum: 1
        targetReps:
          type: integer
          minimum: 1
        targetWeightKg:
          type: number
          format: double
          minimum: 0
          nullable: true

    Template:
      type: object
      required: [id, name, createdAt, exercises]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/TemplateExerciseItem'

    TemplateSummary:
      type: object
      required: [id, name, exerciseCount]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        notes:
          type: string
          nullable: true
        exerciseCount:
          type: integer

    TemplateExerciseInput:
      type: object
      required: [exerciseId]
      description: Position is implied by array order (1-based).
      properties:
        exerciseId:
          type: integer
          format: int64
        targetSets:
          type: integer
          minimum: 1
          default: 3
        targetReps:
          type: integer
          minimum: 1
          default: 10
        targetWeightKg:
          type: number
          format: double
          minimum: 0
          nullable: true

    CreateTemplateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        notes:
          type: string
          nullable: true
          maxLength: 1000
        exercises:
          type: array
          description: Ordered exercise list. Omitted or empty means a template with no exercises.
          items:
            $ref: '#/components/schemas/TemplateExerciseInput'